Add files via upload
This is .xsl file that is required to be present in the same folder/repository etc. as the .xml files which are intended to be viewed online. The file itself was acquired from the "module3-wranglingdata-master" folder, which had been downloaded from the "craftingdigitalhistory" GitHub repository (https://github.com/craftingdigitalhistory/module3-wranglingdata). Note: unfortunately, due to differing security settings, the files cannot be viewed in Chrome. Any wishing to gain access to the encoded work must open the files in either Internet Explorer or Firefox.